[Update] Playground prototype
Did some TLC on the movement Test scene to fix some small bugs.

Created the "Playground" as a copy of the movement test for people to mess around with different Ideas without the worry of breaking too much.

Todo: Reticle placement either projected onto surfaces or as a 2D sprite along the screen as a "sticker"