[Added] Enemy AI
Added a base enemy AI pyramid

made the stationary enemy class and implementation in the enemy testing scene.
Can switch between FORWARD and ATPLAYER to select targeting styles.
FORWARD takes the objects actual forward vector and sets the Weapons origin to that direction.
ATPLAYER takes the same origin transform and rotates it to face the player when a attack is called.